requirements.txt
audioread==2.1.8
certifi==2020.4.5.1
cffi==1.14.0
cycler==0.10.0
decorator==4.4.2
dill==0.3.1.1
future==0.18.2
joblib==0.14.1
kiwisolver==1.2.0
librosa==0.7.2
llvmlite==0.32.0
matplotlib==3.2.1
multiprocess==0.70.9
numba==0.49.0
numpy==1.18.4
Pillow==7.1.2
protobuf==3.11.3
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
resampy==0.2.2
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
SoundFile==0.10.3.post1
tensorboardX==2.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.0
yacs==0.1.7
